transcription/requirements.txt
assemblyai>=0.35.0
python-dotenv>=1.0.0
